backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
pydantic[email]==2.10.3
pydantic-settings==2.7.0
python-multipart==0.0.18
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
celery==5.4.0
redis==5.2.1
boto3==1.35.81
httpx==0.28.1
python-dotenv==1.0.1
pandas==2.2.3
openpyxl==3.1.5
